Add child when dojo is created so that founder have the right to have a social life and children, too :o)
